// ProcMgt/ClientApp/src/app/models/approvalscreen.ts
export class Approvalscreen {
  itemRequestID: string;
  requestTitle: string;
  categoryID: string;
  itemID: string;
  itemTypeID: string;
  makeID: string;
  modelID: string;
  priorityID: string;
  requiredDate: string;
  isReplaceble: boolean;
  assetCode: string;
  noOfUnits: number;
  approverID: string;
  remark: string;
  attachment: string;
  isApproved: boolean;
  isRejected: boolean;
  approvedDateTime: string;
  updatedRequestedDateTime: string;
  approvalComment: string;
  requestedDateTime: string;
  requestedUserID: string;
  departmentID: string;
  approvedUserID: string;
  isIssued: boolean;
  balancedQty: number;
  receivedQty: number;


  categoryName: string; 
  itemDescription: string;
  makeName: string;
  priorityLevelName: string;
  itemTypeName: string;
  modelName: string;
  approverName: string;
  requestedUserName: string;
  approvedUserName: string;
  departmentName: string;


 
}